refactor(agents): extract execution-budget accounting from cook.rs into cook_budget
cook.rs is a 3575-line god-file mixing candidate adoption, execution,
promotion, finalization, baseline materialization, and budget math. First
carve: lift the self-contained execution-budget accounting into a new
cook_budget module.

Moves ExecutionBudgetUsage + execution_budget_usage / budget_remaining /
reserve_remediation_budget — pure functions over aggregate/budget data with no
I/O, which is why they lift cleanly. provider_rotation_attempts (shared with
the executor-identity cluster) is made pub(crate) and referenced from the new
module via super::cook.

Pure move, no behavior change: the fns keep pub(crate) visibility and are
re-exported through mod.rs's pub use, so crate::agent_task_service::* call
sites (including tests.rs) resolve unchanged. cook.rs -97 lines.

Verified: homeboy-agents --lib clean; the budget test
(cook_usage_reads_scheduler_rotation_metadata_and_decrements_budget) passes.
